Initial Commit
I had some commits before this one, but do to some messup of mine I got
to the point where it was easier to delete and recreate .git/ than to do
some rebasing magic that I only partially understand. That's why this
initial commit is already a decently functional version of the software.